pyproject.toml
[project]
name = "ai-agent-product-support"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.8",
    "langchain-google-genai>=3.1.0",
    "langgraph>=1.0.3",
    "numpy>=2.3.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.0",
    "sentence-transformers>=5.1.2",
]
